+++ src/lib/sqs/sqs_producer.ts
@@ -1,37 +1,33 @@
-import { SQSClient, ReceiveMessageCommand } from "@aws-sdk/client-sqs";
+import { SQSClient, SendMessageCommand } from "@aws-sdk/client-sqs";
 import { config } from "../../config";
-import { logger } from "../util/logger";
+import axios from "axios";
 
-interface CrawlerRequest {
+interface CrawlerInput {
   character: string;
   realm: string;
 }
 
 /**
- * reads request from the SQS queue. It will return
- * a list of requests that have been made for the crawler
- * to collect character data
+ * sends a message to the SQS queue used by the crawler lambda
+ * to collect character data from warmane asynchronously
  */
-export async function readCrawlerRequests(): Promise<CrawlerRequest[]> {
-  return [];
-}
-
-export async function receiveSqsMessage(): Promise<CrawlerRequest[]> {
-  const client = new SQSClient({});
-  const command = new ReceiveMessageCommand({
-    QueueUrl: config.crawlerSqsUrl,
-  });
-  const result = await client.send(command);
-  if (result.Messages) {
-    return result.Messages.map((m) => {
-      if (!m.Body) {
-        logger.warn("message with empty body");
-        return;
-      }
-      return JSON.parse(m.Body);
-    }).filter(Boolean);
+export async function requestCrawl(input: CrawlerInput) {
+  if (process.env.AWS_EXECUTION_ENV !== undefined) {
+    await sendSqsMessage(input);
+  } else {
+    await sendLocalMessage(input);
   }
-  return [];
 }
 
-// async function receiveLocalMessage() {}
+async function sendSqsMessage(message: object) {
+  const client = new SQSClient({});
+  const m = new SendMessageCommand({
+    QueueUrl: config.crawlerSqsUrl,
+    MessageBody: JSON.stringify(message),
+  });
+  await client.send(m);
+}
+
+async function sendLocalMessage(message: object) {
+  axios.post(config.crawlerSqsUrl, message);
+}
